Refactoring / simplifying
- Remove project/workspace concept, keep just analysis
- Refactor test into multiple test with pytest.fixture
- Push `--zettle` out of model, into CLI helper (just rename)
