+++ jove/templates/README.md
@@ -1,3 +1,3 @@
-# {title}
+# {name}
 
 Add your analysis notes here.
